requirements.txt
absl-py==1.0.0
attrs==21.4.0
audiomentations==0.20.0
audioread==2.1.8
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cycler==0.11.0
decorator==5.1.0
essentia==2.1b6.dev184
ffmpeg-python==0.2.0
fire==0.2.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.25.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.3.5
matplotlib==3.1.3
musdb==0.4.0
numba==0.48.0
numpy==1.17.4
oauthlib==3.1.1
pandas==1.0.1
Pillow==9.0.1
protobuf==3.19.1
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyloudnorm==0.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytomlpp==1.0.10
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
smmap==5.0.0
SoundFile==0.10.3.post1
sox==1.3.7
spock-config==2.4.2
stempeg==0.2.3
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.1.0
termcolor==1.1.0
torch==1.2.0
torch-tb-profiler==0.3.1
torchaudio==0.3.0
tqdm==4.40.0
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
